requirements.txt
PyPDF2==3.0.1
streamlit==1.22.0
langchain==0.0.132
openai==0.27.6
tiktoken==0.4.0
pypdf==3.8.1
pinecone-client==2.2.1